ParkingLot_Detector/requirement.txt
# requirements.txt 

torch==2.7.0
torchvision==0.22.0
opencv-python==4.11.0.86
easyocr==1.7.2
pytesseract==0.3.13
scikit-image==0.25.2
pillow==11.2.1
numpy==2.1.1
firebase-admin==6.8.0
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
requests==2.32.3
tqdm==4.67.1
PyYAML==6.0.2
python-dateutil==2.9.0.post0
typing_extensions==4.13.2
